Moved class shared by sascalc into sascalc
